Unify BaseManagerFetchAll and BaseManager
