+++ src/application/NodeAllocationManager.ts
@@ -1,14 +1,15 @@
 import type { PteroApp } from '.';
+import { BaseManager } from '../structures/BaseManager';
 import { Allocation } from '../common/app';
 import { Dict } from '../structures/Dict';
-import { FetchOptions, Include } from '../common';
+import { FetchOptions, Include, PaginationMeta } from '../common';
 import caseConv from '../util/caseConv';
 import endpoints from './endpoints';
-import { BaseManagerFetchAll } from '../structures/BaseManagerFetchAll';
 
-export class NodeAllocationManager extends BaseManagerFetchAll {
+export class NodeAllocationManager extends BaseManager {
     public client: PteroApp;
     public cache: Dict<number, Dict<number, Allocation>>;
+    public meta: PaginationMeta;
 
     /** Allowed filter arguments for allocations. */
     get FILTERS() {
