+++ src/application/NodeLocationManager.ts
@@ -1,4 +1,5 @@
 import type { PteroApp } from '.';
+import { BaseManager } from '../structures/BaseManager';
 import { Dict } from '../structures/Dict';
 import {
     FetchOptions,
@@ -13,11 +14,11 @@ import {
 import { ValidationError } from '../structures/Errors';
 import caseConv from '../util/caseConv';
 import endpoints from './endpoints';
-import { BaseManagerFetchAll } from '../structures/BaseManagerFetchAll';
 
-export class NodeLocationManager extends BaseManagerFetchAll {
+export class NodeLocationManager extends BaseManager {
     public client: PteroApp;
     public cache: Dict<number, NodeLocation>;
+    public meta: PaginationMeta;
 
     /**
      * Allowed filter arguments for locations:
