+++ src/application/NodeManager.ts
@@ -1,4 +1,5 @@
 import type { PteroApp } from '.';
+import { BaseManager } from '../structures/BaseManager';
 import { Dict } from '../structures/Dict';
 import { Node } from '../structures/Node';
 import {
@@ -18,11 +19,11 @@ import {
 import { ValidationError } from '../structures/Errors';
 import caseConv from '../util/caseConv';
 import endpoints from './endpoints';
-import { BaseManagerFetchAll } from '../structures/BaseManagerFetchAll';
 
-export class NodeManager extends BaseManagerFetchAll {
+export class NodeManager extends BaseManager {
     public client: PteroApp;
     public cache: Dict<number, Node>;
+    public meta: PaginationMeta;
 
     /**
      * Allowed filter arguments for nodes:
